benchmmark_construction/requirements.txt
mysql-connector-python>=8.0.30
psycopg2-binary==2.9.9
hydra-core==1.3.2
omegaconf==2.3.0
openai==2.7.2
python-dotenv==1.0.0
click==8.2.1